Remove chooseBackground logic from dom.js
