+++ lib/cjs/utilities/chooseWeather.d.ts
@@ -1,0 +1,1 @@
+export declare let chooseWeatherBackground: (weatherCode: number, isDay: number) => void;
